+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 - LoadTest.Scenario.maximumTime() removed.
 - LoadTest.Scenario.Builder introduced along with LoadTest.Performer, AbstractScenario removed, related refactoring.
 - LoadTest.Scenario, couple of scenario() factory methods added.
+- LoadTest.title() renamed name(), related refactoring.
 ### is.codion.common.rmi
 - SerializationWhitelist, now allows arrays but checks the component type.
 - SerializationWhitelist.DryRun, now excludes arrays and includes the component type.